working on view.json

+++ src/main/java/org/vniizht/suburbsweb/model/transformation/level3/tables/Times.java
@@ -14,7 +14,7 @@ import java.sql.Date;
 public class Times {
     private Date            date;
     private String          time;
-    private int             time2;
+    private double          time2;
     private String          dann;
     private String          oper;
     private Date            date_zap;
