+++ src/main/java/org/vniizht/suburbsweb/model/transformation/level3/tables_co22/T4.java
@@ -11,14 +11,14 @@ import javax.persistence.Table;
 @Getter
 @Setter
 public class T4 {
-    private float           yymm;
+    private int             yymm;
     private String          p1;
     private String          p2;
     private int             p3;
     private short           p4;
     private String          p5;
     private String          p6;
-    private float           p7;
-    private float           p8;
+    private long            p7;
+    private long            p8;
     private short           p9;
 }
